refactor(mf): calculator - scss into css
