--- micro-frontends/src/mf-calculator/src/ui/modules/operation-button/operation-button.module.scss
@@ -1,34 +1,0 @@
-.button {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	@include font('medium', 11, 11);
-	border: none;
-	border-radius: 0;
-	background-color: var(--background-color);
-	color: var(--text-color);
-	box-shadow: var(--box-shadow-raised);
-	width: 35px;
-	height: 30px;
-
-	@include hover {
-		cursor: pointer;
-	}
-
-	&:disabled {
-		pointer-events: none;
-		opacity: 0.5;
-	}
-
-	&:not(:disabled):active {
-		box-shadow: var(--box-shadow-sunken);
-	}
-}
-
-.primary {
-	color: $cl-blue00;
-}
-
-.secondary {
-	color: $cl-red;
-}
