+++ micro-frontends/src/mf-calculator/src/ui/modules/operation-button/operation-button.tsx
@@ -1,5 +1,6 @@
+import { getClassNames } from "@windows98/toolkit"
 import type { FunctionComponent } from "react"
-import styles from "./operation-button.module.scss"
+import styles from "./operation-button.module.css"
 import type { OperationButtonProps } from "./operation-button.types"
 
 export const OperationButton: FunctionComponent<OperationButtonProps> = ({
@@ -9,7 +10,7 @@ export const OperationButton: FunctionComponent<OperationButtonProps> = ({
 }) => {
 	return (
 		<button
-			className={`${styles.button} ${styles[variant]}`}
+			className={getClassNames([styles.button, styles[variant]])}
 			data-testid={`mf-calculator-operation-button-${id}`}
 			onClick={onClick}
 			type="button"
