+++ micro-frontends/src/mf-calculator/src/ui/modules/screen/screen.module.css
@@ -1,0 +1,11 @@
+.screen {
+	background-color: var(--ds-background-color-secondary);
+	color: var(--ds-text-color-primary);
+	width: 100%;
+	height: 20px;
+	padding: var(--ds-spacing-sm);
+	text-align: right;
+	box-shadow: var(--ds-border-field);
+	font-family: var(--ds-font-sans-serif-medium);
+	font-size: var(--ds-font-size-xs);
+}
