--- micro-frontends/src/mf-calculator/src/ui/modules/screen/screen.module.scss
@@ -1,10 +1,0 @@
-.screen {
-  background-color: var(--background-color-white);
-  color: var(--text-color);
-  width: 100%;
-  height: 20px;
-  padding: 5px;
-  text-align: right;
-  box-shadow: var(--border-field);
-  @include font('medium', 11, 13);
-}
