+++ micro-frontends/src/mf-calculator/src/ui/modules/screen/screen.tsx
@@ -1,6 +1,7 @@
+import { getClassNames } from "@windows98/toolkit"
 import { observer } from "mobx-react-lite"
 import type { FunctionComponent } from "react"
-import styles from "./screen.module.scss"
+import styles from "./screen.module.css"
 import { useScreen } from "./use-screen.hook"
 
 export const Screen: FunctionComponent = observer(() => {
@@ -8,7 +9,7 @@ export const Screen: FunctionComponent = observer(() => {
 
 	return (
 		<span
-			className={styles.screen}
+			className={getClassNames([styles.screen])}
 			data-testid="mf-calculator-screen"
 		>
 			{value ?? "ERROR"}
