--- micro-frontends/src/mf-calculator/src/ui/modules/wrapper/wrapper.module.scss
@@ -1,6 +1,0 @@
-.wrapper {
-  display: grid;
-  row-gap: 15px;
-  padding: 13px 22px 10px;
-  background-color: var(--background-color);
-}
