+++ micro-frontends/src/mf-calculator/src/ui/modules/wrapper/wrapper.tsx
@@ -1,14 +1,15 @@
+import { getClassNames } from "@windows98/toolkit"
 import type { FunctionComponent } from "react"
 import { Keypad } from "../keypad"
 import { Modal } from "../modal"
 import { Screen } from "../screen"
-import styles from "./wrapper.module.scss"
+import styles from "./wrapper.module.css"
 
 export const Wrapper: FunctionComponent = () => {
 	return (
 		<Modal>
 			<div
-				className={styles.wrapper}
+				className={getClassNames([styles.wrapper])}
 				data-testid="mf-calculator-wrapper"
 			>
 				<Screen />
